composer, deployer がインストール済みのときは再度インストールしない

+++ recipes/default.rb
@@ -29,17 +29,23 @@ end
 
 include_recipe 'php::ini'
 
-execute "install composer and deployer" do
+execute "install composer" do
     user "root"
     command <<-EOC
         php -r "copy('https://getcomposer.org/installer', 'composer-setup.php');"
         php -r "if (hash_file('SHA384', 'composer-setup.php') === '544e09ee996cdf60ece3804abc52599c22b1f40f4323403c44d44fdfdd586475ca9813a858088ffbc1f233e9b180f061') { echo 'Installer verified'; } else { echo 'Installer corrupt'; unlink('composer-setup.php'); } echo PHP_EOL;"
         php composer-setup.php --install-dir=/usr/local/bin --filename=composer
         php -r "unlink('composer-setup.php');"
-        composer require deployer/deployer
+    EOC
+    not_if { File.exists?('/usr/local/bin/composer') }
+end
 
+execute "install deployer" do
+    user "root"
+    command <<-EOC
         curl -LO https://deployer.org/deployer.phar
         mv deployer.phar /usr/local/bin/dep
         chmod +x /usr/local/bin/dep
     EOC
+    not_if { File.exists?('/usr/local/bin/dep') }
 end
